Phase 11 complete, before performance optimization

+++ src/@FEM_Preprocessor_v2/fuseNodes.m
@@ -3,7 +3,7 @@ fprintf('[Mesher] Fusing duplicate nodes (Tol: %g)...\n', tol);
 old_nodes = obj.Mesh.Nodes;
 [unique_nodes, ~, idx_map] = uniquetol(old_nodes, tol, 'ByRows', true);
 obj.Mesh.Nodes = unique_nodes;
-% Use reshape to ensure the matrix dimensions are preserved
-obj.Mesh.Elements = reshape(idx_map(obj.Mesh.Elements), size(obj.Mesh.Elements));
+% Use reshape to ensure matrix dimensions, cast back to int32
+obj.Mesh.Elements = int32(reshape(idx_map(obj.Mesh.Elements), size(obj.Mesh.Elements)));
 fprintf(' -> Reduced from %d to %d nodes.\n', size(old_nodes,1), size(unique_nodes,1));
 end
